% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimodularity.R
\name{allTotallyUnimodular}
\alias{allTotallyUnimodular}
\title{Test if a list of matrices are all unimodular}
\usage{
allTotallyUnimodular(L)
}
\arguments{
\item{L}{A list of objects of class matrix.}
}
\value{
logical vector of length \code{length(L)}
}
\description{
Helper function for \code{\link{raghavachari}}
}
\seealso{
\code{\link{is_totally_unimodular}}
}
\keyword{internal}

