% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substval.R
\name{subst_value}
\alias{subst_value}
\title{Substitute a value in a system of linear (in)equations}
\usage{
subst_value(A, b, variables, values, remove_columns = FALSE, eps = 1e-08)
}
\arguments{
\item{A}{\code{[numeric]} matrix}

\item{b}{\code{[numeric]} vector}

\item{variables}{\code{[numeric|logical|character]} vector of column indices in \code{A}}

\item{values}{\code{[numeric]} vecor of values to substitute.}

\item{remove_columns}{\code{[logical]} Remove spurious columns when substituting?}

\item{eps}{\code{[numeric]} scalar. Any value with absolute value below \code{eps} will be 
interpreted as zero.}
}
\value{
A \code{list} with the following components:

\itemize{
  \item{\code{A}: the \code{A} corresponding to the simplified sytem.}
  \item{\code{b}: the constant vector corresponding to the new system}
}
}
\description{
Substitute a value in a system of linear (in)equations
}
\section{Details}{

A system of the form \code{Ax <= b} can be simplified if one or more of the
\code{x[i]} values is fixed.
}

